requirements.txt
flask
Flask-SQLAlchemy 
flask-cors 
cryptography 
PyJWT
PyMySQL 
mysql-connector-python
python-dotenv 
SQLAlchemy 
Werkzeug 
